dcp-815 separate process_link.py
