+++ tests/exporter/graph/test_experiment_graph.py
@@ -1,6 +1,7 @@
 from unittest import TestCase
 
-from exporter.graph.experiment_graph import LinkSet, ProcessLink, SupplementaryFileLink, SupplementedEntity
+from exporter.graph.experiment_graph import LinkSet, SupplementaryFileLink, SupplementedEntity
+from exporter.graph.process_link import ProcessLink
 
 
 class TestUtils:
